reuse coursePageSection component
