+++ components/textbook-beta/CoursePagesSection.vue
@@ -49,8 +49,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import { Component } from 'vue-property-decorator'
-import { textbookBetaIntroductionCourseStartLearningUrl } from '~/constants/appLinks'
+import { Component, Prop } from 'vue-property-decorator'
 import { CtaClickedEventProp } from '~/constants/segment'
 
 interface Course {
@@ -62,58 +61,10 @@ interface Course {
 
 @Component
 export default class CoursePagesSection extends Vue {
-  activeCourseLabel = '';
+  @Prop({ type: Array, required: true }) courses!: Course[]
+  @Prop({ type: String, required: true }) imgBase!: string
 
-  courses: Course[] = [
-    {
-      image: 'why-quantum-computing-preview.png',
-      label: 'Why quantum computing?',
-      segment: { cta: 'why-quantum-computing', location: 'course' },
-      url: textbookBetaIntroductionCourseStartLearningUrl
-    },
-    {
-      image: 'the-atoms-of-computation-preview.png',
-      label: 'The atoms of computation',
-      segment: { cta: 'the-atoms-of-computation', location: 'course' },
-      url: 'https://learn.qiskit.org/course/introduction/the-atoms-of-computation'
-    },
-    {
-      image: 'what-is-quantum-preview.png',
-      label: 'What is quantum?',
-      segment: { cta: 'what-is-quantum', location: 'course' },
-      url: 'https://learn.qiskit.org/course/introduction/what-is-quantum'
-    },
-    {
-      image: 'describing-quantum-computers-preview.png',
-      label: 'Describing quantum computers',
-      segment: { cta: 'describing-quantum-computers', location: 'course' },
-      url: 'https://learn.qiskit.org/course/introduction/describing-quantum-computers'
-    },
-    {
-      image: 'entangled-states-preview.png',
-      label: 'Entangled states',
-      segment: { cta: 'entangled-states', location: 'course' },
-      url: 'https://learn.qiskit.org/course/introduction/entangled-states'
-    },
-    {
-      image: 'visualizing-entanglement-preview.png',
-      label: 'Visualizing Entanglement',
-      segment: { cta: 'visualizing-entanglement-preview', location: 'course' },
-      url: 'https://learn.qiskit.org/course/introduction/visualizing-entanglement'
-    },
-    {
-      image: 'grovers-search-algorithm-preview.png',
-      label: 'Grover’s search algorithm',
-      segment: { cta: 'grovers-search-algorithm', location: 'course' },
-      url: 'https://learn.qiskit.org/course/introduction/grovers-search-algorithm'
-    },
-    {
-      image: 'project-preview.png',
-      label: 'Project',
-      segment: { cta: 'project', location: 'course' },
-      url: 'https://learn.qiskit.org/course/introduction/project'
-    }
-  ]
+  activeCourseLabel = '';
 
   get activeCourse () {
     const activeCourse = this.courses.find(
@@ -127,8 +78,7 @@ export default class CoursePagesSection extends Vue {
       return ''
     }
 
-    const imageUrlBase =
-      '/images/textbook-beta/course/introduction-course'
+    const imageUrlBase = this.imgBase
 
     return `${imageUrlBase}/${this.activeCourse.image}`
   }
