+++ pages/textbook-beta/course/introduction-course.vue
@@ -19,8 +19,8 @@
       </template>
     </AppPageHeaderWithImage>
     <PrerequisiteMaterialSection class="introduction-course-page__section" :data="introductionCoursePrerequisiteData" />
-    <LinkListSection class="introduction-course-page__section" :links="links" :title="recommendedReadingsSectionTitle" />
-    <CoursePagesSection class="introduction-course-page__section" />
+    <LinkListSection class="introduction-course-page__section" :links="links" />
+    <CoursePagesSection class="introduction-course-page__section" :courses="courses" :img-base="imageUrlBase" />
   </main>
 </template>
 
@@ -28,6 +28,14 @@
 import { Component } from 'vue-property-decorator'
 import QiskitPage from '~/components/logic/QiskitPage.vue'
 import { GeneralLink, textbookBetaIntroductionCourseStartLearningUrl, introductionCoursePrerequisites } from '~/constants/appLinks'
+import { CtaClickedEventProp } from '~/constants/segment'
+
+interface Course {
+  image: string;
+  label: string;
+  segment: CtaClickedEventProp;
+  url: string;
+}
 
 @Component({
   layout: 'default-max',
@@ -91,7 +99,58 @@ export default class IntroductionCoursePage extends QiskitPage {
     }
   ]
 
-  recommendedReadingsSectionTitle = 'External recommended readings'
+  courses: Course[] = [
+    {
+      image: 'why-quantum-computing-preview.png',
+      label: 'Why quantum computing?',
+      segment: { cta: 'why-quantum-computing', location: 'course' },
+      url: textbookBetaIntroductionCourseStartLearningUrl
+    },
+    {
+      image: 'the-atoms-of-computation-preview.png',
+      label: 'The atoms of computation',
+      segment: { cta: 'the-atoms-of-computation', location: 'course' },
+      url: 'https://learn.qiskit.org/course/introduction/the-atoms-of-computation'
+    },
+    {
+      image: 'what-is-quantum-preview.png',
+      label: 'What is quantum?',
+      segment: { cta: 'what-is-quantum', location: 'course' },
+      url: 'https://learn.qiskit.org/course/introduction/what-is-quantum'
+    },
+    {
+      image: 'describing-quantum-computers-preview.png',
+      label: 'Describing quantum computers',
+      segment: { cta: 'describing-quantum-computers', location: 'course' },
+      url: 'https://learn.qiskit.org/course/introduction/describing-quantum-computers'
+    },
+    {
+      image: 'entangled-states-preview.png',
+      label: 'Entangled states',
+      segment: { cta: 'entangled-states', location: 'course' },
+      url: 'https://learn.qiskit.org/course/introduction/entangled-states'
+    },
+    {
+      image: 'visualizing-entanglement-preview.png',
+      label: 'Visualizing Entanglement',
+      segment: { cta: 'visualizing-entanglement-preview', location: 'course' },
+      url: 'https://learn.qiskit.org/course/introduction/visualizing-entanglement'
+    },
+    {
+      image: 'grovers-search-algorithm-preview.png',
+      label: 'Grover’s search algorithm',
+      segment: { cta: 'grovers-search-algorithm', location: 'course' },
+      url: 'https://learn.qiskit.org/course/introduction/grovers-search-algorithm'
+    },
+    {
+      image: 'project-preview.png',
+      label: 'Project',
+      segment: { cta: 'project', location: 'course' },
+      url: 'https://learn.qiskit.org/course/introduction/project'
+    }
+  ]
+
+  imageUrlBase = '/images/textbook-beta/course/introduction-course'
 }
 </script>
